Move TM2 ToNumeric() to util package
